+++ packages/ts/src/plugin.ts
@@ -35,6 +35,7 @@ import arrayTypes from "./rules/arrayTypes.ts";
 import arrayUnnecessaryLengthChecks from "./rules/arrayUnnecessaryLengthChecks.ts";
 import asConstAssertions from "./rules/asConstAssertions.ts";
 import assignmentOperatorShorthands from "./rules/assignmentOperatorShorthands.ts";
+import asyncFunctionAwaits from "./rules/asyncFunctionAwaits.ts";
 import asyncPromiseExecutors from "./rules/asyncPromiseExecutors.ts";
 import asyncUnnecessaryPromiseWrappers from "./rules/asyncUnnecessaryPromiseWrappers.ts";
 import atAccesses from "./rules/atAccesses.ts";
@@ -250,6 +251,7 @@ import regexZeroQuantifiers from "./rules/regexZeroQuantifiers.ts";
 import requireImports from "./rules/requireImports.ts";
 import responseJsonMethods from "./rules/responseJsonMethods.ts";
 import restrictedIdentifiers from "./rules/restrictedIdentifiers.ts";
+import restrictedImports from "./rules/restrictedImports.ts";
 import returnAssignments from "./rules/returnAssignments.ts";
 import returnThisTypes from "./rules/returnThisTypes.ts";
 import selfAssignments from "./rules/selfAssignments.ts";
@@ -346,6 +348,7 @@ export const ts = createPlugin({
 		arrayUnnecessaryLengthChecks,
 		asConstAssertions,
 		assignmentOperatorShorthands,
+		asyncFunctionAwaits,
 		asyncPromiseExecutors,
 		asyncUnnecessaryPromiseWrappers,
 		atAccesses,
@@ -562,6 +565,7 @@ export const ts = createPlugin({
 		requireImports,
 		responseJsonMethods,
 		restrictedIdentifiers,
+		restrictedImports,
 		returnAssignments,
 		returnThisTypes,
 		selfAssignments,