+++ packages/vitest/src/rules/testCasesWithinDescribes.ts
@@ -61,8 +61,7 @@ export default ruleCreator.createRule(typescriptLanguage, {
 					}
 				},
 				"CallExpression:exit": (node) => {
-					// eslint-disable-next-line @typescript-eslint/no-non-null-assertion
-					switch (parseVitestFunctionCall(node!)?.name) {
+					switch (parseVitestFunctionCall(node)?.name) {
 						case "describe":
 							insideDescribeStack -= 1;
 							break;